requirements.txt
python-dotenv==1.0.0
openai==1.68.2
httpx==0.27.2
gradio==5.23.1
numpy==2.2.4
faiss-cpu==1.10.0
langchain==0.3.21
sentence-transformers==4.0.1
transformers==4.50.1
pypdf==5.4.0